+++ backend/src/navigatorhandlers.py
